feat(plugin): add test case for noderesourcesfit

+++ plugin/noderesourcesfit.go
@@ -97,11 +97,12 @@ func (n *NodeResourcesFit) leastRequestedScore(requested, capacity int64) int64 
 	if capacity == 0 {
 		return 0
 	}
+
 	if requested > capacity {
 		return 0
 	}
 
-	return ((capacity - requested) * int64(proto.MaxNodeScore)) / capacity
+	return ((capacity - requested) * proto.MaxNodeScore) / capacity
 }
 
 // nolint:typecheck
